Reorganized strategy pattern to make it neater
